Percentage

+++ alcohol.scala
@@ -78,7 +78,7 @@ def sorted_country_consumption() : List[(String, Long)] = {
   val list = for {
     nameA <- process_alcs(get_csv_page(url_alcohol))
     nameB <- process_pops(get_csv_file(file_population))
-    total = nameA._2 * nameB._2
+    total = nameA._2.toLong * nameB._2.toLong
 
     if (nameA._1 == nameB._1) } yield (nameA._1, total.toLong)
